Changelog: http://itextpdf.com/changelog/557
Release Notes: http://itextpdf.com/release/iText557
Download: https://github.com/itext/itextpdf/releases/tag/5.5.7

+++ itext/src/main/java/com/itextpdf/text/pdf/codec/TIFFFaxDecoder.java
@@ -56,7 +56,7 @@ public class TIFFFaxDecoder {
     private int bitPointer, bytePointer;
     private byte[] data;
     private int w, h;
-    private int fillOrder;
+    private long fillOrder;
     
     // Data structures needed to store changing elements for the previous
     // and the current scanline
@@ -588,7 +588,7 @@ public class TIFFFaxDecoder {
      * @param w
      * @param h
      */
-    public TIFFFaxDecoder(int fillOrder, int w, int h) {
+    public TIFFFaxDecoder(long fillOrder, int w, int h) {
         this.fillOrder = fillOrder;
         this.w = w;
         this.h = h;
